UK: Standardised workflow and popolo
Put the UK data through our normal workflow, as if the data were coming
from morph. (It can’t atm, because the parlparse JSON is too large, and
blows the memory limits there). This will make it simpler to merge with
other sources (e.g. YNMP and Wikidata), and ensure that the Popolo
output is consistent with other countries.