New merge process for Wikidata
Rather than stepping through all the Wikidata data and trying to match
it to known people (which has lots of false positives), step through
the people we care about, and try to find an appropriate Wikidata
record.

Also allow an ‘overrides’  hash to short-circuit this where the fuzzy
matching gets it wrong.