Bulgaria: don't fetch name__en
Not every record will have a `name__en` field, so fetching it means
they will end up with empty values. We already coalesce with name__bg
(which will always be set) to make a plain `name` field anyway. If we
really want a multilingual:en name record, that's something we should
fix globally.