Feed all person data into the reconciliation template
Rather than just feeding the fields specified in instructions.json into
the template we instead give it the whole existing and incoming objects
so that any information can be displayed on that page.

This also changes the reconciliation csv format so there are now just
two columns, 'id' and 'uuid'.