Add temporary code to convert legacy reconciliation CSVs
This is just a temporary measure. I wanted to put it into a separate
script that we could run as a one off, but I need access to the merged
rows and the easiest place to get to those is in the middle of the
combine_sources file.

This currently only works correctly for existing reconciliation files
which have an id in the header. Reconciliation files where the incoming
data doesn't have an id are going to be a bit more troublesome.